+++ src/ansible_navigator/cli.py
@@ -1,7 +1,6 @@
 # cspell:ignore getpid, gmtime, msecs
 """start here
 """
-import datetime
 import logging
 import os
 import signal
@@ -11,7 +10,6 @@ from copy import deepcopy
 from curses import wrapper
 from pathlib import Path
 from typing import List
-from typing import Optional
 from typing import Union
 
 from pkg_resources import working_set
@@ -28,7 +26,7 @@ from .configuration_subsystem import NavigatorConfiguration
 from .image_manager import ImagePuller
 from .initialization import error_and_exit_early
 from .initialization import parse_and_update
-from .utils.compatibility import zoneinfo
+from .logging import setup_logger
 from .utils.functions import ExitMessage
 from .utils.functions import ExitPrefix
 from .utils.functions import LogMessage
@@ -75,68 +73,6 @@ def pull_image(args):
         image_puller.pull_stdout()
     if image_puller.assessment.exit_messages:
         error_and_exit_early(image_puller.assessment.exit_messages)
-
-
-class Formatter(logging.Formatter):
-    """Format a logging timestamp using a time zone."""
-
-    def __init__(self, *args, **kwargs):
-        """Initialize the logging formatter.
-
-        :param args: The arguments
-        :param kwargs: The keyword arguments
-        """
-        self._time_zone = kwargs.pop("time_zone")
-        super().__init__(*args, **kwargs)
-
-    def formatTime(self, record: logging.LogRecord, _datefmt: Optional[str] = None) -> str:
-        """Format the log timestamp.
-
-        :param record: The log record
-        :param _datefmt: The optiona date format
-        :returns: The timestamp
-        """
-        if self._time_zone == "local":
-            return (
-                datetime.datetime.fromtimestamp(record.created, datetime.timezone.utc)
-                .astimezone()
-                .isoformat()
-            )
-        return datetime.datetime.fromtimestamp(
-            record.created,
-            tz=zoneinfo.ZoneInfo(self._time_zone),
-        ).isoformat()
-
-
-def setup_logger(args: ApplicationConfiguration) -> None:
-    """set up the logger
-
-    :param args: The CLI args
-    """
-    if os.path.exists(args.log_file) and args.log_append is False:
-        os.remove(args.log_file)
-    handler = logging.FileHandler(args.log_file)
-
-    time_zone = args.entry("time_zone").value.current
-    # When the configuration is rolled back, the time_zone will be C.NOT_SET
-    if isinstance(args.time_zone, Constants):
-        time_zone = args.entry("time_zone").value.default
-
-    formatter = Formatter(
-        fmt="%(asctime)s %(levelname)s '%(name)s.%(funcName)s' %(message)s",
-        time_zone=time_zone,
-    )
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
-    log_level = getattr(logging, args.log_level.upper())
-    logger.setLevel(log_level)
-    logger.info("New %s instance, logging initialized", APP_NAME)
-
-    # set ansible-runner logs
-    runner_logger = logging.getLogger("ansible-runner")
-    runner_logger.setLevel(log_level)
-    runner_logger.addHandler(handler)
-    logger.info("New ansible-runner instance, logging initialized")
 
 
 def run(args: ApplicationConfiguration) -> ActionReturn:
